+++ app/models/post.ts
@@ -1,4 +1,5 @@
 import Model, { attr } from '@ember-data/model';
+import { memberAction } from 'ember-api-actions';
 import User from './user';
 
 export default class Post extends Model {
@@ -43,6 +44,11 @@ export default class Post extends Model {
 
   @attr({ defaultValue: false })
   disableEmojis?: boolean;
+
+  /**
+   * Gets a modified version of the post's message used for quoting.
+   */
+  quote = memberAction<undefined, PostQuote>({ path: 'quote', type: 'GET' });
 }
 
 // DO NOT DELETE: this is how TypeScript knows how to look up your models.
@@ -63,4 +69,8 @@ export interface PostPreview {
   date: string;
   threadId: string;
   boardId: string;
+}
+
+export interface PostQuote {
+  message: string;
 }